FossilOrigin-Name: bf55a4220a74022e8be5f8a0cd7f4ca395a36e1666e14113e7e60c4a94cf0b3a

+++ ext/wasm/mkdist.sh
@@ -193,8 +193,9 @@ done
 #
 # Done copying files. Now zip it up...
 #
-svi=${dirTop}/version-info
+svi=./version-info
 vnum=$($svi --download-version)
+[ "" = "$vnum" ] && die "version number is empty!"
 vdir=${buildName}-${vnum}
 fzip=${vdir}.zip
 rm -fr ${vdir} ${fzip}
